Add gitlab e2e
