+++ gitlab/tests/conftest.py
@@ -22,7 +22,7 @@ from .common import (
 )
 
 
-@pytest.fixture(scope="session", autouse=True)
+@pytest.fixture(scope="session")
 def dd_environment():
     """
     Spin up and initialize gitlab
