+++ gitlab/tox.ini
@@ -6,6 +6,8 @@ envlist =
     bench
 
 [testenv]
+description=
+    py{27,37}: e2e ready
 dd_check_style = true
 usedevelop = true
 platform = linux|darwin|win32
@@ -13,8 +15,8 @@ passenv =
     DOCKER*
     COMPOSE*
 deps =
-    -e../datadog_checks_base[deps]
-    -rrequirements-dev.txt
+    -e ../datadog_checks_base[deps]
+    -r requirements-dev.txt
 commands =
     pip install -r requirements.in
     pytest -v {posargs} --benchmark-skip
